Represent no relationship data different than null data
Like `Document`, `Relationship` needs to differentiate between
`"data":null` in the JSON and no data key at all, so have
`Relationship.t` default to `:unset` instead of `nil`, the same way
`Document.t` works now.  This means adding a Poison.Encoder
implementation to not encode the `:unset` and changing the `ToParams`
behaviour to allow for an `{:error, :unset}` return that the
`Relationship.to_params` can return when `data` is `:unset`, so that
`Relationships.to_params` will skip including the output in the map of
all relationship params.